Сортировка и ошибка

+++ FakeNFT/Services/Requests/RequestConstants.swift
@@ -8,7 +8,7 @@ enum RequestConstants {
     static let application = "application/x-www-form-urlencoded"
     static let nftToken = "20a4069e-6e51-4477-894a-e6fdc9a4bb95"
     static let nftHeader = "X-Practicum-Mobile-Token"
-    static let timeoutInterval = Double(5) // seconds
+    static let timeoutInterval = Double(60) // seconds
     static let baseApiPath = "/api/v1"
 }
 // TODO: разобраться с дубликатами
